+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The complete changelog for the Costs to Expect REST API, follows the format defi
 - New route /resource-types/[resource-type]/items.
 - `include-categories` and `include-subcategories` parameters for /resource-types/[resource-type]/items route.
 - New route /summary/resource-types/[resource-type]/items.
+- Added `resources`, `years`, `year`, `months`, `month`, `categories`, `category`, `subcategories` and `subcategory` parameters to /summary/resource-types/[resource-type]/items route. 
 
 ### Changed
 - Updated the API landing page to point to the new website.