+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The complete changelog for the Costs to Expect REST API, follows the format defi
 ## [v1.13.0] - 2019-xx-xx
 ### Added 
 - New route /resource-types/[resource-type]/items.
-- `include-categories` and `include-subcategories` parameters for /resource-types/[resource-type]/items route.
+- Added `include-categories` and `include-subcategories` parameters to /resource-types/[resource-type]/items route.
+- Added `year`, `month`, `category` and `subcategory` parameters to /resource-types/[resource-type]/items route.
 - New route /summary/resource-types/[resource-type]/items.
-- Added `resources`, `years`, `year`, `months`, `month`, `categories`, `category`, `subcategories` and `subcategory` parameters to /summary/resource-types/[resource-type]/items route. 
+- Added `resources`, `years`, `year`, `months`, `month`, `categories`, `category`, `subcategories` and `subcategory` parameters to /summary/resource-types/[resource-type]/items route.  
 
 ### Changed
 - Updated the API landing page to point to the new website.